requirements.txt
cffi==1.8.3
cryptography==1.5.2
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.10.2
django-countries==4.0
django-jsonfield==1.0.1
django-localflavor==1.3
enum34==1.1.6
httplib2==0.9.2
idna==2.1
ipaddress==1.0.17
ndg-httpsclient==0.4.2
oauthlib==2.0.0
psycopg2==2.6.2
pyasn1==0.1.9
pycparser==2.16
PyJWT==1.4.2
pyOpenSSL==16.2.0
python-openid==2.2.5
python-social-auth==0.2.21
requests==2.11.1
requests-oauthlib==0.7.0
six==1.10.0
static3==0.7.0
stripe==1.41.1
http://oxd.gluu.org/
